Added Instrumentation and Unit test
Added Input Validation tests and Imroved Modularity of code
