+++ offix/src/androidTest/java/org/aerogear/offix/ExampleInstrumentedTest.java
@@ -22,5 +22,6 @@ public class ExampleInstrumentedTest {
         Context appContext = InstrumentationRegistry.getTargetContext();
 
         assertEquals("org.aerogear.offixsdk.test", appContext.getPackageName());
+
     }
 }
