+++ sample/build.gradle
@@ -24,6 +24,7 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
 }
 
 dependencies {
@@ -57,8 +58,11 @@ dependencies {
     implementation "android.arch.work:work-runtime-ktx:1.0.1"
     implementation 'com.google.code.gson:gson:2.8.5'
 
+    androidTestImplementation 'org.hamcrest:hamcrest-library:1.3'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
+    testImplementation 'org.mockito:mockito-core:2.7.22'
+    androidTestImplementation 'org.mockito:mockito-android:2.7.22'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 
     /*
