+++ sample/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
             android:supportsRtl="true"
             android:usesCleartextTraffic="true"
             android:theme="@style/AppTheme">
-        <activity android:name="org.aerogear.graphqlandroid.activities.MainActivity"
+        <activity android:name=".activities.MainActivity"
                   android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
