+++ sample/src/main/java/org/aerogear/graphqlandroid/activities/MainActivity.kt
@@ -8,36 +8,31 @@ import android.text.Editable
 import android.text.TextWatcher
 import android.util.Log
 import android.view.LayoutInflater
-import android.widget.Toast
 import com.apollographql.apollo.ApolloCall
-import com.apollographql.apollo.ApolloQueryWatcher
 import com.apollographql.apollo.api.Response
 import com.apollographql.apollo.exception.ApolloException
 import com.apollographql.apollo.fetcher.ApolloResponseFetchers
-import com.apollographql.apollo.rx2.Rx2Apollo
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.schedulers.Schedulers
-import io.reactivex.subscribers.DisposableSubscriber
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.alertfrag_create_tasks.view.*
-import org.aerogear.graphqlandroid.*
+import org.aerogear.graphqlandroid.FindAllTasksQuery
+import org.aerogear.graphqlandroid.R
+import org.aerogear.graphqlandroid.Utils
+
+import org.aerogear.graphqlandroid.managers.MutationManager.createTask;
+import org.aerogear.graphqlandroid.managers.SubscriptionManager.subscriptionNewTask
+import org.aerogear.graphqlandroid.managers.SubscriptionManager.subscriptionNewUser
+import org.aerogear.graphqlandroid.managers.SubscriptionManager.subscriptionUpdateTask
+import org.aerogear.graphqlandroid.managers.SubscriptionManager.subscriptionUpdateUser
 import org.aerogear.graphqlandroid.adapter.TaskAdapter
 import org.aerogear.graphqlandroid.model.UserOutput
-import org.aerogear.graphqlandroid.type.TaskInput
-import org.aerogear.graphqlandroid.type.UserInput
 import org.aerogear.offix.Offline
 
 class MainActivity : AppCompatActivity() {
-
     var tasksList = arrayListOf<UserOutput>()
     private val TAG = javaClass.simpleName
     val taskAdapter by lazy {
         TaskAdapter(tasksList, this)
     }
-
-    private val disposables = CompositeDisposable()
-    private var apolloQueryWatcher: ApolloQueryWatcher<FindAllTasksQuery.Data>? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -86,7 +81,7 @@ class MainActivity : AppCompatActivity() {
                         ilTitleTask.error = resources.getString(R.string.task_title_error)
                     } else {
                         val desc = inflatedView.etDescTask.text.toString()
-                        createTask(title, desc)
+                        createTask(title, desc, this)
                         customAlert.dismiss()
                     }
                 }
@@ -109,7 +104,6 @@ class MainActivity : AppCompatActivity() {
         tasksList.clear()
         getTasks()
     }
-
     private fun getTasks() {
         Log.e(TAG, " ----- getTasks")
 
@@ -167,431 +161,14 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    fun updateTask(id: String, title: String, description: String) {
-        Log.e(TAG, "inside update title in MainActivity")
-
-        /*
-        As version is assumed to be auto incremented ( //TODO Have to make changes in sqlite db)
-        */
-        val input =
-            TaskInput.builder().title(title).version(1).description(description).status("test")
-                .build()
-
-        val mutation = UpdateTaskMutation.builder().id(id).input(input).build()
-
-        Log.e(TAG, " updateTask ********: - $mutation")
-
-        val mutationCall = Utils.getApolloClient(this)?.mutate(
-            mutation
-        )?.refetchQueries(apolloQueryWatcher?.operation()?.name())
-
-        Log.e(TAG, " updateTask class name: - ${mutation.javaClass.simpleName}")
-        Log.e(
-            TAG,
-            " updateTask 20: - ${mutationCall?.requestHeaders(com.apollographql.apollo.request.RequestHeaders.builder().build())}"
-        )
-        Log.e(TAG, " updateTask 21: - ${mutationCall?.operation()?.queryDocument()}")
-        Log.e(TAG, " updateTask 22: - ${mutationCall?.operation()?.variables()?.valueMap()}")
-        Log.e(TAG, " updateTask 23: - ${mutationCall?.operation()?.name()}")
-
-        val callback = object : ApolloCall.Callback<UpdateTaskMutation.Data>() {
-            override fun onFailure(e: ApolloException) {
-                Log.e("onFailure() updateTask", "${mutation.variables().valueMap()}")
-                e.printStackTrace()
-            }
-
-            override fun onResponse(response: Response<UpdateTaskMutation.Data>) {
-                val result = response.data()?.updateTask()
-
-                //In case of conflicts data returned from the server id null.
-                result?.let {
-                    Log.e(TAG, "onResponse-UpdateTask- $it")
-                }
-            }
-        }
-        mutationCall?.enqueue(callback)
-        if (Offline.isNetwork()) {
-            Toast.makeText(this@MainActivity, "Task with id $id is updated", Toast.LENGTH_LONG)
-                .show()
-        } else {
-            Toast.makeText(
-                this@MainActivity,
-                "Task with id $id is stored offline. Changes will be synced to the server when app comes online.",
-                Toast.LENGTH_LONG
-            ).show()
-
-        }
-    }
-
-    fun updateUser(
-        idUser: String,
-        taskId: String,
-        title: String,
-        firstName: String,
-        lastName: String,
-        email: String
-    ) {
-        Log.e(TAG, "inside updateUser in MainActivity")
-
-        val input =
-            UserInput.builder().title(title).lastName(lastName).firstName(firstName).email(email)
-                .taskId(taskId)
-                .creationmetadataId(taskId).build()
-
-        val mutation = UpdateUserMutation.builder().id(idUser).input(input).build()
-
-        Log.e(TAG, " updateUser ********: - $mutation")
-
-        val mutationCall = Utils.getApolloClient(this)?.mutate(
-            mutation
-        )?.refetchQueries(apolloQueryWatcher?.operation()?.name())
-
-        Log.e(TAG, " updateUser class name: - ${mutation.javaClass.simpleName}")
-        Log.e(
-            TAG,
-            " updateUser 20: - ${mutationCall?.requestHeaders(com.apollographql.apollo.request.RequestHeaders.builder().build())}"
-        )
-        Log.e(TAG, " updateUser 21: - ${mutationCall?.operation()?.queryDocument()}")
-        Log.e(TAG, " updateUser 22: - ${mutationCall?.operation()?.variables()?.valueMap()}")
-        Log.e(TAG, " updateUser 23: - ${mutationCall?.operation()?.name()}")
-
-        val callback = object : ApolloCall.Callback<UpdateUserMutation.Data>() {
-            override fun onFailure(e: ApolloException) {
-                Log.e("onFailure() updateTask", "${mutation.variables().valueMap()}")
-                e.printStackTrace()
-            }
-
-            override fun onResponse(response: Response<UpdateUserMutation.Data>) {
-                val result = response.data()?.updateUser()
-
-                //In case of conflicts data returned from the server id null.
-                result?.let {
-                    Log.e(TAG, "onResponse-UpdateTask- $it")
-                }
-            }
-        }
-        mutationCall?.enqueue(callback)
-        if (Offline.isNetwork()) {
-            Toast.makeText(
-                this,
-                resources.getText(R.string.user_updated, idUser),
-                Toast.LENGTH_LONG
-            ).show()
-        } else {
-            Toast.makeText(
-                this,
-                resources.getText(R.string.user_updates_stored, idUser),
-                Toast.LENGTH_LONG
-            ).show()
-        }
-
-    }
-
-    private fun createTask(title: String, description: String) {
-        Log.e(TAG, "inside create title")
-
-        /*
-         As version is assumed to be auto incremented ( //TODO Have to make changes in sqlite db)
-         */
-        val input =
-            TaskInput.builder().title(title).description(description).version(1).status("test")
-                .build()
-
-        val mutation = CreateTaskMutation.builder().input(input).build()
-
-        val mutationCall = Utils.getApolloClient(this)?.mutate(
-            mutation
-        )?.refetchQueries(apolloQueryWatcher?.operation()?.name())
-
-        Log.e(TAG, " updateTask 22: - ${mutationCall?.operation()?.variables()?.valueMap()}")
-
-
-        val callback = object : ApolloCall.Callback<CreateTaskMutation.Data>() {
-            override fun onFailure(e: ApolloException) {
-                Log.e("onFailure() updateTask", "${mutation.variables().valueMap()}")
-                e.printStackTrace()
-            }
-
-            override fun onResponse(response: Response<CreateTaskMutation.Data>) {
-                val result = response.data()?.createTask()
-
-                //In case of conflicts data returned from the server id null.
-                result?.let {
-                    Log.e(TAG, "onResponse-UpdateTask- $it")
-                }
-            }
-        }
-        mutationCall?.enqueue(callback)
-
-        if (Offline.isNetwork()) {
-            Toast.makeText(this, "Mutation with title $title is created", Toast.LENGTH_LONG).show()
-        } else {
-            Toast.makeText(
-                this,
-                "Create mutation with title $title is stored offline. Changes will be synced to the server when app comes online.",
-                Toast.LENGTH_LONG
-            ).show()
-        }
-    }
-
-    fun createUser(
-        title: String,
-        firstName: String,
-        lastName: String,
-        email: String,
-        taskId: String
-    ) {
-        Log.e(TAG, "inside create user")
-
-        val input =
-            UserInput.builder().taskId(taskId).email(email).firstName(firstName).lastName(lastName)
-                .title(title)
-                .creationmetadataId(taskId).build()
-
-        val mutation = CreateUserMutation.builder().input(input).build()
-
-        val mutationCall = Utils.getApolloClient(this)?.mutate(
-            mutation
-        )?.refetchQueries(apolloQueryWatcher?.operation()?.name())
-
-        Log.e(TAG, " createUser 22: - ${mutationCall?.operation()?.variables()?.valueMap()}")
-        val callback = object : ApolloCall.Callback<CreateUserMutation.Data>() {
-            override fun onFailure(e: ApolloException) {
-                Log.e("onFailure() updateTask", "${mutation.variables().valueMap()}")
-                e.printStackTrace()
-            }
-
-            override fun onResponse(response: Response<CreateUserMutation.Data>) {
-                val result = response.data()?.createUser()
-
-                //In case of conflicts data returned from the server id null.
-                result?.let {
-                    Log.e(TAG, "onResponse-UpdateTask- $it")
-                }
-            }
-        }
-        mutationCall?.enqueue(callback)
-        if (Offline.isNetwork()) {
-            Toast.makeText(
-                this,
-                "Task with id $taskId is assigned to $firstName $lastName",
-                Toast.LENGTH_LONG
-            )
-                .show()
-        } else {
-            Toast.makeText(
-                this,
-                "Update in user where Task with id $taskId is assigned to $firstName $lastName is stored offline. Changes will be synced to the server when app comes online.",
-                Toast.LENGTH_LONG
-            ).show()
-        }
-    }
-
-    fun checkAndUpdateTask(id: String, title: String, description: String) {
-        Log.e(TAG, "inside checkAndUpdateTask in MainActivity")
-
-        /*
-        As version is assumed to be auto incremented ( //TODO Have to make changes in sqlite db)
-        */
-
-        val mutation =
-            CheckAndUpdateTaskMutation.builder().version(1).id(id).title(title)
-                .description(description).status("test")
-                .build()
-
-        Log.e(TAG, " checkAndUpdateTask ********: - $mutation")
-
-        val mutationCall = Utils.getApolloClient(this)?.mutate(
-            mutation
-        )?.refetchQueries(apolloQueryWatcher?.operation()?.name())
-
-        Log.e(TAG, " checkAndUpdateTask class name: - ${mutation.javaClass.simpleName}")
-        Log.e(
-            TAG,
-            " checkAndUpdateTask 20: - ${mutationCall?.requestHeaders(com.apollographql.apollo.request.RequestHeaders.builder().build())}"
-        )
-        Log.e(TAG, " checkAndUpdateTask 21: - ${mutationCall?.operation()?.queryDocument()}")
-        Log.e(
-            TAG,
-            " checkAndUpdateTask 22: - ${mutationCall?.operation()?.variables()?.valueMap()}"
-        )
-        Log.e(TAG, " checkAndUpdateTask 23: - ${mutationCall?.operation()?.name()}")
-
-        val callback = object : ApolloCall.Callback<CheckAndUpdateTaskMutation.Data>() {
-            override fun onFailure(e: ApolloException) {
-                Log.e("onFailure() updateTask", "${mutation.variables().valueMap()}")
-                e.printStackTrace()
-            }
-
-            override fun onResponse(response: Response<CheckAndUpdateTaskMutation.Data>) {
-                val result = response.data()?.checkAndUpdateTask()
-
-                //In case of conflicts data returned from the server id null.
-                result?.let {
-                    Log.e(TAG, "onResponse-UpdateTask- $it")
-                }
-            }
-        }
-        mutationCall?.enqueue(callback)
-        if (Offline.isNetwork()) {
-            Toast.makeText(
-                this@MainActivity,
-                "Task with id $id is updated without any Conflict.",
-                Toast.LENGTH_LONG
-            )
-                .show()
-        } else {
-            Toast.makeText(
-                this@MainActivity,
-                "Task with id $id is stored offline. Changes will be synced to the server when app comes online.",
-                Toast.LENGTH_LONG
-            ).show()
-        }
-    }
-
-    private fun subscriptionNewTask() {
-        val subscription = NewTaskSubscription()
-        val subscriptionCall = Utils.getApolloClient(this)
-            ?.subscribe(subscription)
-
-        disposables.add(
-            Rx2Apollo.from<NewTaskSubscription.Data>(subscriptionCall!!)
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribeWith(
-                    object : DisposableSubscriber<Response<NewTaskSubscription.Data>>() {
-                        override fun onNext(response: Response<NewTaskSubscription.Data>) {
-                            val res = response.data()?.newTask()
-                            res?.let {
-                            }
-                        }
-
-                        override fun onError(e: Throwable) {
-                            Log.e(TAG, e.message, e)
-                        }
-
-                        override fun onComplete() {
-                            Log.e(TAG, "Subscription new task added exhausted")
-                        }
-                    }
-                )
-        )
-    }
-
-    private fun subscriptionUpdateTask() {
-
-        val subscription = UpdatedTaskSubscription()
-        val subscriptionCall = Utils.getApolloClient(this)
-            ?.subscribe(subscription)
-
-        disposables.add(Rx2Apollo.from<UpdatedTaskSubscription.Data>(subscriptionCall!!)
-            .subscribeOn(Schedulers.io())
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribeWith(
-                object : DisposableSubscriber<Response<UpdatedTaskSubscription.Data>>() {
-                    override fun onNext(response: Response<UpdatedTaskSubscription.Data>) {
-                        val res = response.data()?.updatedTask()
-                        res?.let {
-                            Log.e(
-                                TAG,
-                                " inside subscriptionUpdateTask ${it.title()} mutated upon updating"
-                            )
-                        }
-                    }
-
-                    override fun onError(e: Throwable) {
-                        Log.e(TAG, e.message, e)
-                    }
-
-                    override fun onComplete() {
-                        Log.e(TAG, "subscriptionUpdateTask exhausted")
-                    }
-                }
-            )
-        )
-    }
-
-    private fun subscriptionNewUser() {
-        val subscription = NewUserSubscription()
-        val subscriptionCall = Utils.getApolloClient(this)
-            ?.subscribe(subscription)
-
-        disposables.add(
-            Rx2Apollo.from<NewUserSubscription.Data>(subscriptionCall!!)
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribeWith(
-                    object : DisposableSubscriber<Response<NewUserSubscription.Data>>() {
-                        override fun onNext(response: Response<NewUserSubscription.Data>) {
-                            val res = response.data()?.newUser()
-                            res?.let {
-                                Log.e(
-                                    TAG,
-                                    " inside subscriptionNewUser ${it.title()} mutated upon new title"
-                                )
-                            }
-                        }
-
-                        override fun onError(e: Throwable) {
-                            Log.e(TAG, e.message, e)
-                        }
-
-                        override fun onComplete() {
-                            Log.e(TAG, "Subscription new user added exhausted")
-                        }
-                    }
-                )
-        )
-    }
-
-    private fun subscriptionUpdateUser() {
-
-        val subscription = UpdatedUserSubscription()
-        val subscriptionCall = Utils.getApolloClient(this)
-            ?.subscribe(subscription)
-
-        disposables.add(Rx2Apollo.from<UpdatedUserSubscription.Data>(subscriptionCall!!)
-            .subscribeOn(Schedulers.io())
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribeWith(
-                object : DisposableSubscriber<Response<UpdatedUserSubscription.Data>>() {
-                    override fun onNext(response: Response<UpdatedUserSubscription.Data>) {
-
-                        val res = response.data()?.updatedUser()
-                        res?.let {
-                            Log.e(
-                                TAG,
-                                " inside subscriptionUpdateUser ${it.title()} mutated upon updating"
-                            )
-                        }
-                    }
-
-                    override fun onError(e: Throwable) {
-                        Log.e(TAG, e.message, e)
-                    }
-
-                    override fun onComplete() {
-                        Log.e(TAG, "subscriptionUpdateUser exhausted")
-                    }
-                }
-            )
-        )
-    }
-
     override fun onStart() {
         if (Offline.isNetwork()) {
-            subscriptionNewTask()
-            subscriptionNewUser()
-            subscriptionUpdateTask()
-            subscriptionUpdateUser()
+            subscriptionNewTask(this)
+            subscriptionNewUser(this)
+            subscriptionUpdateTask(this)
+            subscriptionUpdateUser(this)
         }
         super.onStart()
-    }
-
-    override fun onDestroy() {
-        apolloQueryWatcher?.cancel()
-        disposables.dispose()
-        super.onDestroy()
     }
 }
